Improved styles v2
1. Add style "ip".
2. Save/load paragraph properties as floats.
3. Edit paragraph properties as floats with precision of one digit.
https://github.com/bibledit/cloud/issues/990

+++ stylesv2/logic.h
@@ -135,11 +135,11 @@ struct Paragraph {
   TwoState underline{};
   TwoState smallcaps{};
   TextAlignment text_alignment{};
-  int space_before{};
-  int space_after{};
-  int left_margin{};
-  int right_margin{};
-  int first_line_indent{};
+  float space_before{};
+  float space_after{};
+  float left_margin{};
+  float right_margin{};
+  float first_line_indent{};
 };
 
 std::ostream& operator<<(std::ostream& os, const Paragraph paragraph);
